Check valid URIs in a TimeMap without resorting to regex.
A previous attempt per #289 caused URI-Rs with a port number to
fail the regex. This was problematic when a URI-M specific a
URI-R with a port number, which was uncommon but exhibited with
captures from @phonedude's from a long time ago. Changed to
use the URL construct built into JS and created a function to
check this and return on exception, which indicates not a valid
URI.

Closes #289