Fall back to mkdc memgator instance when ODUCS Memgator instance is down
Added a not-so-elegant health check function and a fairly naive
fallback algorithm that will still continually do a HEAD request
on the ODUCS aggregator. A better approach might be to only do this
periodically instead of each request. Regardless, this is a
sufficient initial implementation to make the tool more resilient.

Fixed #201
Fixed #284